+++ Controller/Payment/RestoreCart.php
@@ -1,0 +1,60 @@
+<?php
+/**
+ * Copyright 2026 Vipps
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
+ * TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
+ */
+declare(strict_types=1);
+
+namespace Vipps\Payment\Controller\Payment;
+
+use Magento\Framework\App\ActionInterface;
+use Magento\Framework\App\CsrfAwareActionInterface;
+use Magento\Framework\App\Request\InvalidRequestException;
+use Magento\Framework\App\RequestInterface;
+use Magento\Framework\Controller\ResultFactory;
+use Vipps\Payment\Model\Express\CartRestorer;
+use Vipps\Payment\Model\Express\CartRestoreCookie;
+
+class RestoreCart implements ActionInterface, CsrfAwareActionInterface
+{
+    public function __construct(
+        private readonly ResultFactory $resultFactory,
+        private readonly CartRestoreCookie $cartRestoreCookie,
+        private readonly CartRestorer $cartRestorer
+    ) {
+    }
+
+    public function execute()
+    {
+        $result = $this->resultFactory->create(ResultFactory::TYPE_JSON);
+
+        $quoteId = $this->cartRestoreCookie->getPending();
+        if (!$quoteId) {
+            return $result->setData(['restored' => false]);
+        }
+
+        $this->cartRestoreCookie->clearPending();
+
+        return $result->setData(['restored' => $this->cartRestorer->restore($quoteId)]);
+    }
+
+    public function createCsrfValidationException(RequestInterface $request): ?InvalidRequestException
+    {
+        return null;
+    }
+
+    public function validateForCsrf(RequestInterface $request): ?bool
+    {
+        return true;
+    }
+}